refactor demo components
